api/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
fastapi==0.115.12
flake8==7.2.0
h11==0.16.0
httptools==0.6.4
idna==3.10
kombu==5.5.4
mccabe==0.7.0
motor==3.7.1
packaging==25.0
prompt_toolkit==3.0.51
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
